requirements.txt
click>=8.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
rich>=13.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
weasyprint>=60.0
jinja2>=3.1.0
